# api-doc.yml
swagger: "2.0"
info:
  description: ""
  version: "0.1.0"
  title: "CMRS 接口描述文档"
  termsOfService: "http://117.78.6.206:9580"
  license:
    name: "Apache 2.0"
    url: "http://www.apache.org/licenses/LICENSE-2.0.html"
host: "139.159.162.133:9580"
basePath: "/"
...
schemes:
  - http
security:
  - CMRSSecurity: []
consumes:
  - application/json
produces:
  - application/json
paths:
...
  /update-self-password:
    post:
      tags:
        - "管理员维护自己信息"
      summary: "管理员更新自己的密码"
      parameters:
        - in: body
          name: "管理员更新自己的密码命令"
          required: true
          schema:
            $ref: "#/definitions/UpdateUserPasswordCommand"
      responses:
        204:
          description: "退出登录成功"
...
  /set-roles-to-user/:userId:
    post:
      tags:
        - "管理员维护用户信息"
      summary: "管理员为用户设置角色"
      parameters:
        - name: "userId"
          type: string
          in: path
          required: true
        - in: body
          name: "设置角色命令"
          description: "超级管理员或者用户管理员为用户设置角色"
          required: true
          schema:
            $ref: "#/definitions/SetRolesToUserCommand"
      responses:
        204:
          description: "分配角色成功"
        400:
          headers:
            X-BIZ-ERR-CODE:
              type: string
          description: "请求数据异常"
          schema:
            $ref: "#/definitions/EX40001"
  /update-details-of-user/:userId:
    post:
      tags:
        - "管理员维护用户信息"
      summary: "管理员更新用户信息"
      parameters:
        - name: "userId"
          type: string
          required: true
          in: path
        - in: body
          name: "更新用户信息命令"
          description: "超级管理员或者用户管理员更新用户信息"
          required: true
          schema:
            $ref: '#/definitions/UpdateUserDetailsCommand'
      responses:
        204:
          description: "更新成功"
        404:
          description: "指定的用户不存在"
  /update-user/:userId/password/:passwordId:
    post:
      tags:
        - "管理员维护用户信息"
      summary: "管理员更新用户密码"
      parameters:
        - name: "userId"
          type: string
          in: path
          required: true
        - name: "userPasswordId"
          type: string
          in: path
          required: true
        - in: body
          name: "更新用户信息命令"
          description: "超级管理员或者用户管理员新用户密码"
          required: true
          schema:
            $ref: '#/definitions/UpdateUserPasswordCommand'
      responses:
        204:
          description: "更新成功"
        404:
          description: "指定的用户不存在"
  /create-clerk:
    post:
      tags:
        - "管理员维护用户信息"
      summary: "管理员创建职员接口"
      parameters:
        - in: body
          name: "管理员创建职员命令"
          required: true
          schema:
            $ref: "#/definitions/CreateClerkWithUsernameCommand"
      responses:
        200:
          description: "创建职员成功"
          schema:
            $ref: "#/definitions/AggregateCreatedDTO"
  /get-user-details/:id:
    get:
      tags:
        - "管理员维护用户信息"
      summary: "管理员查询用户详情接口"
      parameters:
        - name: "id"
          in: path
          required: true
          type: string
          description: "要查询的用户ID"
      responses:
        200:
          description: "查询成功"
          schema:
            $ref: '#/definitions/UserDetails'
  /get-user-list:
    get:
      tags:
        - "管理员维护用户信息"
      summary: "管理员查询用户信息列表接口"
      parameters:
        - name: "page"
          type: integer
          in: query
          description: "页码"
        - name: "size"
          type: integer
          in: query
          description: "页面最大记录数"
      responses:
        200:
          description: "查询成功"
          schema:
            type: object
            properties:
              content:
                type: array
                items:
                  $ref: "#/definitions/UserDetails"
              pageInfo:
                $ref: "#/definitions/PageInfo"
...
  /update-self-details:
    post:
      tags:
        - "用户维护自己信息"
      summary: "用户（所有角色）更新个人信息接口"
      parameters:
        - name: "更新用户自己的信息"
          in: body
          required: true
          schema:
            $ref: "#/definitions/UpdateUserSelfDetailsDTO"
      responses:
        204:
          description: "更新成功"
...
  /find-user:
    get:
      tags:
        - "通用"
      parameters:
        - name: userId
          type: string
          in: query
        - name: realName
          type: string
          minLength: 2
          in: query
      responses:
        200:
          description: "查询成功"
          schema:
            $ref: "#/definitions/FindUserResultDTO"
...
securityDefinitions:
  CMRSSecurity:
    type: apiKey
    in: header
    name: Authorization
definitions:
...
  CreateClerkWithUsernameCommand:
    type: object
    properties:
      username:
        type: string
        example: "zhangsan"
      realName:
        type: string
        example: "张三"
      password:
        type: string
        example: "babd4a84418fb04015ac773f7459727c"
    required:
      - username
      - realName
      - password
  SetRolesToUserCommand:
    type: object
    properties:
      roleIdList:
        type: array
        items:
          type: string
          example:
            - "a71ed7a7ee2cf1555a77038538c2c68c"
            - "t54327a7ee2cf1555a77038538c2c68c"
    required:
      - userId
      - roleId
  UpdateUserDetailsCommand:
    type: object
    properties:
      realName:
        type: string
        example: "李四"
    required:
      - id
      - realName
  UpdateUserPasswordCommand:
    type: object
    properties:
      newUserPassword:
        type: string
        description: "密码md5哈希后的字符串"
        example: "babd4a84418fb04015ac773f7459727c"
    required:
      - newUserPassword
...
  AggregateCreatedDTO:
    type: object
    properties:
      code:
        type: string
        example: "20000"
      data:
        type: object
        properties:
          id:
            type: string
            example: "b6bdf1c1-5094-45d8-ad1d-231e0f19b48b"
  UserDetails:
    type: object
    properties:
      id:
        type: string
        example: "2f83252d-ab8b-4568-ad12-7a758f6e9f43"
      realName:
        type: string
        example: "超级管理员"
      accountViewList:
        type: array
        items:
          $ref: "#/definitions/AccountView"
      roleViewList:
        type: array
        items:
          type: object
          properties:
            id:
              type: string
              example: "9fdd4a72bdfe15c619db8e526968da25"
            name:
              type: string
              example: "ROLE_SUPER_ADMIN"
            createdAt:
              type: string
              example: "2019-07-12T14:32:41.358+0000"
      userSessionId:
        type: string
        description: "会话ID，和Token相同"
        example: "cbf6416f-f055-484b-aeda-bb501bde67ee"
      userSessionExpiredAt:
        type: string
        example: "2019-07-21T15:15:33.482+0000"
      userPasswordId:
        type: string
        example: "e988ad63-5472-4688-b2d8-c67964e6ec77"
      createdAt:
        type: string
        example: "2019-07-12T14:32:41.358+0000"
      updatedAt:
        type: string
        example: null
  AccountView:
    type: object
    properties:
      id:
        type: string
        example: "7a0376b3219c4d1d5447c70cccde4f9e"
      idInType:
        type: string
        example: "admin"
      type:
        type: string
        example: "USERNAME"
      createdAt:
        type: string
        example: "2019-06-17T23:22:00.255+0000"
      updatedAt:
        type: string
        example: "null"
...
  PageInfo:
    type: object
    properties:
      totalElements:
        type: integer
        example: 5
      totalPages:
        type: integer
        example: 1
      number:
        description: "当前页码"
        type: integer
        example: 0
      size:
        description: "页面最大记录条数"
        type: integer
        example: 20
...
  UpdateUserSelfDetailsDTO:
    type: object
    properties:
      id:
        type: string
        example: "b6bdf1c1-5094-45d8-ad1d-231e0f19b48b"
      realName:
        type: string
        example: "李四"
    required:
      - id
      - realName
...
  FindUserResultDTO:
    type: array
    maxItems: 5
    items:
      type: object
      properties:
        id:
          type: string
        realName:
          type: string
...
  EX40001:
    type: object
    properties:
      message:
        type: string
        example: "指定的用户不存在"
...